Move consensus messages to node crate
