+++ exonum-node/Cargo.toml
@@ -19,9 +19,11 @@ travis-ci = { repository = "exonum/exonum" }
 bit-vec = "0.6.0"
 byteorder = { version = "1.2.7", features = [ "i128" ] }
 bytes = "0.4.11"
+chrono = "0.4.6"
 failure = "0.1.5"
 futures = "0.1.25"
 log = "0.4.6"
+protobuf = { version = "2.8.1", features = ["with-serde"] }
 rand = "0.7"
 serde = "1.0.101"
 serde_derive = "1.0.101"
@@ -36,18 +38,22 @@ tokio-dns-unofficial = "0.4.0"
 
 exonum = { version = "0.13.0-rc.2", path = "../exonum" }
 exonum-api = { version = "0.13.0-rc.2", path = "../components/api" }
+exonum-derive = { version = "0.13.0-rc.2", path = "../components/derive" }
+exonum-merkledb = { version = "0.13.0-rc.2", path = "../components/merkledb" }
+exonum-proto = { version = "0.13.0-rc.2", path = "../components/proto" }
 exonum_sodiumoxide = { version = "0.0.23", optional = true }
 
 [dev-dependencies]
 bincode = "1.2.1"
-chrono = "0.4.6"
 criterion = "0.3.0"
 hex = "0.4.0"
 pretty_assertions = "0.6.1"
+serde_json = "1.0.44"
 
-exonum-derive = { version = "0.13.0-rc.2", path = "../components/derive" }
-exonum-merkledb = { version = "0.13.0-rc.2", path = "../components/merkledb" }
 exonum-rust-runtime = { version = "0.13.0-rc.2", path = "../runtimes/rust" }
+
+[build-dependencies]
+exonum-build = { version = "0.13.0-rc.2", path = "../components/build" }
 
 [features]
 default = ["exonum_sodiumoxide"]
