+++ exonum-node/src/proto/consensus.proto
@@ -18,30 +18,15 @@ syntax = "proto3";
 
 package exonum.consensus;
 
+// TODO: is it needed?
 option java_package = "com.exonum.core.messages";
 
 import "types.proto";
 import "common.proto";
 import "blockchain.proto";
+import "messages.proto";
 import "runtime.proto";
 import "google/protobuf/timestamp.proto";
-
-// Container for the signed messages.
-message SignedMessage {
-  // Payload of the message as a serialized `ExonumMessage`.
-  bytes payload = 1;
-  // Public key of the author of the message.
-  exonum.crypto.PublicKey author = 2;
-  // Digital signature over the payload created with a secret key of the author of the message.
-  exonum.crypto.Signature signature = 3;
-}
-
-// Container for a verified message.
-message Verified {
-  SignedMessage raw = 1;
-}
-
-// List of consensus messages
 
 message Connect {
   string host = 1;
@@ -69,15 +54,6 @@ message Prevote {
   uint32 round = 3;
   exonum.crypto.Hash propose_hash = 4;
   uint32 locked_round = 5;
-}
-
-message Precommit {
-  uint32 validator = 1;
-  uint64 height = 2;
-  uint32 round = 3;
-  exonum.crypto.Hash propose_hash = 4;
-  exonum.crypto.Hash block_hash = 5;
-  google.protobuf.Timestamp time = 6;
 }
 
 message BlockResponse {
@@ -128,7 +104,7 @@ message ExonumMessage {
     Connect connect = 2;
     Status status = 3;
 
-    Precommit precommit = 4;
+    exonum.messages.Precommit precommit = 4;
     Propose propose = 5;
     Prevote prevote = 6;
