+++ exonum-node/src/sandbox/tests/round_details.rs
@@ -23,7 +23,7 @@ use exonum::{
     crypto::Hash,
     helpers::{Height, Round, ValidatorId},
     merkledb::ObjectHash,
-    messages::{PrevotesRequest, TransactionsRequest, Verified},
+    messages::Verified,
 };
 use log::info;
 
@@ -32,7 +32,10 @@ use std::{collections::HashSet, convert::TryFrom, time::Duration};
 use crate::constants::{
     PREVOTES_REQUEST_TIMEOUT, PROPOSE_REQUEST_TIMEOUT, TRANSACTIONS_REQUEST_TIMEOUT,
 };
-use crate::sandbox::{sandbox_tests_helper::*, timestamping_sandbox, timestamping_sandbox_builder};
+use crate::{
+    messages::{PrevotesRequest, TransactionsRequest},
+    sandbox::{sandbox_tests_helper::*, timestamping_sandbox, timestamping_sandbox_builder},
+};
 
 /// check scenario:
 /// HANDLE FULL PROPOSE
