+++ exonum-node/src/sandbox/tests/transactions.rs
@@ -20,17 +20,20 @@ use exonum::{
     crypto::{gen_keypair, Hash},
     helpers::{Height, Round, ValidatorId},
     merkledb::{BinaryValue, ObjectHash},
-    messages::{AnyTx, Verified, TX_RES_EMPTY_SIZE, TX_RES_PB_OVERHEAD_PAYLOAD},
+    messages::{AnyTx, Verified},
 };
 
 use std::time::Duration;
 
-use crate::constants::TRANSACTIONS_REQUEST_TIMEOUT;
 use crate::sandbox::{
     config_updater::TxConfig,
     sandbox_tests_helper::*,
     timestamping::{TimestampingTxGenerator, DATA_SIZE},
     timestamping_sandbox, timestamping_sandbox_builder, Milliseconds, Sandbox,
+};
+use crate::{
+    constants::TRANSACTIONS_REQUEST_TIMEOUT,
+    messages::{TX_RES_EMPTY_SIZE, TX_RES_PB_OVERHEAD_PAYLOAD},
 };
 
 const MAX_PROPOSE_TIMEOUT: Milliseconds = 200;
