+++ exonum/Cargo.toml
@@ -20,7 +20,7 @@ travis-ci = { repository = "exonum/exonum" }
 [dependencies]
 log = "0.4.6"
 hex = "0.4"
-bit-vec = "0.6.0"
+bit-vec = "0.6.0" # ???
 serde = "1.0.101"
 serde_derive = "1.0.101"
 serde_json = "1.0.19"
