+++ exonum/src/messages/mod.rs
@@ -90,326 +90,144 @@
 //!
 //!
 
-pub use self::{signed::Verified, types::*};
-
-use exonum_merkledb::{BinaryValue, ObjectHash};
-
-use std::borrow::Cow;
-
-use crate::{
-    crypto::{Hash, PublicKey, SecretKey, PUBLIC_KEY_LENGTH, SIGNATURE_LENGTH},
-    helpers::{Height, Round, ValidatorId},
+pub use self::{
+    signed::{IntoMessage, Verified},
+    types::*,
 };
 
+use crate::crypto::{PUBLIC_KEY_LENGTH, SIGNATURE_LENGTH};
+
 mod signed;
-#[cfg(test)]
-mod tests;
 mod types;
 
 /// Lower bound on the size of the correct `SignedMessage`.
-/// Size of message fields + protobuf overhead.
+/// This is the size of message fields + protobuf overhead.
 #[doc(hidden)]
 pub const SIGNED_MESSAGE_MIN_SIZE: usize = PUBLIC_KEY_LENGTH + SIGNATURE_LENGTH + 8;
 
-#[doc(hidden)]
-pub const TX_RES_EMPTY_SIZE: usize = SIGNED_MESSAGE_MIN_SIZE + PUBLIC_KEY_LENGTH + 8;
+#[cfg(test)]
+mod tests {
+    use chrono::Utc;
+    use exonum_crypto::{self as crypto, gen_keypair, Signature};
+    use exonum_merkledb::BinaryValue;
+    use exonum_proto::ProtobufConvert;
+    use protobuf::Message;
 
-/// When we add transaction to TransactionResponse message we will add some overhead
-/// to the message size due to protobuf.
-/// This is higher bound on this overhead.
-/// Tx response message size <= TX_RES_EMPTY_SIZE + (tx1 size + TX_RES_PB_OVERHEAD_PAYLOAD) +
-///                             + (tx2 size + TX_RES_PB_OVERHEAD_PAYLOAD) + ...
-#[doc(hidden)]
-pub const TX_RES_PB_OVERHEAD_PAYLOAD: usize = 8;
-
-/// Service messages.
-#[derive(Debug, Clone, PartialEq)]
-pub enum Service {
-    /// Transaction message.
-    AnyTx(Verified<AnyTx>),
-    /// Connect message.
-    Connect(Verified<Connect>),
-    /// Status message.
-    Status(Verified<Status>),
-}
-
-impl Service {
-    fn as_raw(&self) -> &SignedMessage {
-        match self {
-            Service::AnyTx(ref msg) => msg.as_raw(),
-            Service::Connect(ref msg) => msg.as_raw(),
-            Service::Status(ref msg) => msg.as_raw(),
-        }
-    }
-}
-
-/// Consensus messages.
-#[derive(Debug, Clone, PartialEq)]
-pub enum Consensus {
-    /// `Precommit` message.
-    Precommit(Verified<Precommit>),
-    /// `Propose` message.
-    Propose(Verified<Propose>),
-    /// `Prevote` message.
-    Prevote(Verified<Prevote>),
-}
-
-impl Consensus {
-    fn as_raw(&self) -> &SignedMessage {
-        match self {
-            Consensus::Precommit(ref msg) => msg.as_raw(),
-            Consensus::Propose(ref msg) => msg.as_raw(),
-            Consensus::Prevote(ref msg) => msg.as_raw(),
-        }
-    }
-}
-
-/// Response messages.
-#[derive(Debug, Clone, PartialEq)]
-pub enum Responses {
-    /// Transactions response message.
-    TransactionsResponse(Verified<TransactionsResponse>),
-    /// Block response message.
-    BlockResponse(Verified<BlockResponse>),
-}
-
-impl Responses {
-    fn as_raw(&self) -> &SignedMessage {
-        match self {
-            Responses::TransactionsResponse(ref msg) => msg.as_raw(),
-            Responses::BlockResponse(ref msg) => msg.as_raw(),
-        }
-    }
-}
-
-impl From<Verified<TransactionsResponse>> for Responses {
-    fn from(msg: Verified<TransactionsResponse>) -> Self {
-        Responses::TransactionsResponse(msg)
-    }
-}
-
-impl From<Verified<BlockResponse>> for Responses {
-    fn from(msg: Verified<BlockResponse>) -> Self {
-        Responses::BlockResponse(msg)
-    }
-}
-
-/// Request messages.
-#[derive(Debug, Clone, PartialEq)]
-pub enum Requests {
-    /// Propose request message.
-    ProposeRequest(Verified<ProposeRequest>),
-    /// Transactions request message.
-    TransactionsRequest(Verified<TransactionsRequest>),
-    /// Prevotes request message.
-    PrevotesRequest(Verified<PrevotesRequest>),
-    /// Peers request message.
-    PeersRequest(Verified<PeersRequest>),
-    /// Block request message.
-    BlockRequest(Verified<BlockRequest>),
-    /// Request of uncommitted transactions.
-    PoolTransactionsRequest(Verified<PoolTransactionsRequest>),
-}
-
-impl Requests {
-    fn as_raw(&self) -> &SignedMessage {
-        match self {
-            Requests::ProposeRequest(ref msg) => msg.as_raw(),
-            Requests::TransactionsRequest(ref msg) => msg.as_raw(),
-            Requests::PrevotesRequest(ref msg) => msg.as_raw(),
-            Requests::PeersRequest(ref msg) => msg.as_raw(),
-            Requests::BlockRequest(ref msg) => msg.as_raw(),
-            Requests::PoolTransactionsRequest(ref msg) => msg.as_raw(),
-        }
-    }
-}
-
-/// Representation of the Exonum message which is divided into categories.
-#[derive(Debug, Clone, PartialEq)]
-pub enum Message {
-    /// Service messages.
-    Service(Service),
-    /// Consensus messages.
-    Consensus(Consensus),
-    /// Responses messages.
-    Responses(Responses),
-    /// Requests messages.
-    Requests(Requests),
-}
-
-impl Message {
-    /// Creates a new signed Exonum message from the given value.
-    pub fn from_value<T: Into<ExonumMessage>>(
-        message: T,
-        author: PublicKey,
-        secret_key: &SecretKey,
-    ) -> Self {
-        Self::from(Verified::from_value(message.into(), author, secret_key))
-    }
-
-    /// Deserialize message from signed message.
-    pub fn from_signed(signed: SignedMessage) -> Result<Self, failure::Error> {
-        signed.into_verified::<ExonumMessage>().map(From::from)
-    }
-
-    /// Checks buffer and returns instance of `Message`.
-    pub fn from_raw_buffer(buffer: Vec<u8>) -> Result<Self, failure::Error> {
-        SignedMessage::from_bytes(buffer.into()).and_then(Self::from_signed)
-    }
-
-    /// Get inner SignedMessage.
-    pub fn as_raw(&self) -> &SignedMessage {
-        match self {
-            Message::Service(ref msg) => msg.as_raw(),
-            Message::Consensus(ref msg) => msg.as_raw(),
-            Message::Requests(ref msg) => msg.as_raw(),
-            Message::Responses(ref msg) => msg.as_raw(),
-        }
-    }
-}
-
-impl PartialEq<SignedMessage> for Message {
-    fn eq(&self, other: &SignedMessage) -> bool {
-        self.as_raw() == other
-    }
-}
-
-macro_rules! impl_message_from_verified {
-    ( $($concrete:ident: $category:ident),* ) => {
-        $(
-            impl From<Verified<$concrete>> for Message {
-                fn from(msg: Verified<$concrete>) -> Self {
-                    Message::$category($category::$concrete(msg))
-                }
-            }
-
-            impl std::convert::TryFrom<Message> for Verified<$concrete> {
-                type Error = failure::Error;
-
-                fn try_from(msg: Message) -> Result<Self, Self::Error> {
-                    if let Message::$category($category::$concrete(msg)) = msg {
-                        Ok(msg)
-                    } else {
-                        Err(failure::format_err!(
-                            "Given message is not a {}::{}",
-                            stringify!($category),
-                            stringify!($concrete)
-                        ))
-                    }
-                }
-            }
-        )*
-
-        impl From<Verified<ExonumMessage>> for Message {
-            fn from(msg: Verified<ExonumMessage>) -> Self {
-                let raw = msg.raw;
-                match msg.inner {
-                    $(
-                        ExonumMessage::$concrete(inner) => {
-                            let inner = Verified::<$concrete> { raw, inner };
-                            Self::from(inner)
-                        }
-                    )*
-                }
-            }
-        }
+    use super::*;
+    use crate::{
+        helpers::{Height, Round, ValidatorId},
+        proto::schema::messages as proto,
     };
-}
 
-impl_message_from_verified! {
-    AnyTx: Service,
-    Connect: Service,
-    Status: Service,
-    Precommit: Consensus,
-    Prevote: Consensus,
-    Propose: Consensus,
-    BlockResponse: Responses,
-    TransactionsResponse: Responses,
-    BlockRequest: Requests,
-    PeersRequest: Requests,
-    PrevotesRequest: Requests,
-    ProposeRequest: Requests,
-    TransactionsRequest: Requests,
-    PoolTransactionsRequest: Requests
-}
-
-impl Requests {
-    /// Returns public key of the message recipient.
-    pub fn to(&self) -> PublicKey {
-        match *self {
-            Requests::ProposeRequest(ref msg) => msg.payload().to,
-            Requests::TransactionsRequest(ref msg) => msg.payload().to,
-            Requests::PrevotesRequest(ref msg) => msg.payload().to,
-            Requests::PeersRequest(ref msg) => msg.payload().to,
-            Requests::BlockRequest(ref msg) => msg.payload().to,
-            Requests::PoolTransactionsRequest(ref msg) => msg.payload().to,
-        }
+    #[test]
+    fn test_signed_message_min_size() {
+        let (public_key, secret_key) = gen_keypair();
+        let msg = SignedMessage::new(vec![], public_key, &secret_key);
+        assert_eq!(SIGNED_MESSAGE_MIN_SIZE, msg.into_bytes().len())
     }
 
-    /// Returns author public key of the message sender.
-    pub fn author(&self) -> PublicKey {
-        match *self {
-            Requests::ProposeRequest(ref msg) => msg.author(),
-            Requests::TransactionsRequest(ref msg) => msg.author(),
-            Requests::PrevotesRequest(ref msg) => msg.author(),
-            Requests::PeersRequest(ref msg) => msg.author(),
-            Requests::BlockRequest(ref msg) => msg.author(),
-            Requests::PoolTransactionsRequest(ref msg) => msg.author(),
-        }
-    }
-}
+    #[test]
+    fn test_message_roundtrip() {
+        let (pub_key, secret_key) = gen_keypair();
+        let ts = Utc::now();
 
-impl Consensus {
-    /// Returns author public key of the message sender.
-    pub fn author(&self) -> PublicKey {
-        match *self {
-            Consensus::Propose(ref msg) => msg.author(),
-            Consensus::Prevote(ref msg) => msg.author(),
-            Consensus::Precommit(ref msg) => msg.author(),
-        }
-    }
+        let msg = Verified::from_value(
+            Precommit::new(
+                ValidatorId(123),
+                Height(15),
+                Round(25),
+                crypto::hash(&[1, 2, 3]),
+                crypto::hash(&[3, 2, 1]),
+                ts,
+            ),
+            pub_key,
+            &secret_key,
+        );
 
-    /// Returns validator id of the message sender.
-    pub fn validator(&self) -> ValidatorId {
-        match *self {
-            Consensus::Propose(ref msg) => msg.payload().validator(),
-            Consensus::Prevote(ref msg) => msg.payload().validator(),
-            Consensus::Precommit(ref msg) => msg.payload().validator(),
-        }
+        let bytes = msg.to_bytes();
+        let message =
+            SignedMessage::from_bytes(bytes.into()).expect("Cannot deserialize signed message");
+        let msg_roundtrip = message
+            .into_verified::<Precommit>()
+            .expect("Failed to check precommit");
+        assert_eq!(msg, msg_roundtrip);
     }
 
-    /// Returns height of the message.
-    pub fn height(&self) -> Height {
-        match *self {
-            Consensus::Propose(ref msg) => msg.payload().height(),
-            Consensus::Prevote(ref msg) => msg.payload().height(),
-            Consensus::Precommit(ref msg) => msg.payload().height(),
-        }
+    #[test]
+    fn test_signed_message_unusual_protobuf() {
+        let (pub_key, secret_key) = gen_keypair();
+
+        let mut ex_msg = proto::CoreMessage::new();
+        let precommit_msg = Precommit::new(
+            ValidatorId(123),
+            Height(15),
+            Round(25),
+            crypto::hash(&[1, 2, 3]),
+            crypto::hash(&[3, 2, 1]),
+            Utc::now(),
+        );
+        ex_msg.set_precommit(precommit_msg.to_pb());
+        let mut payload = ex_msg.write_to_bytes().unwrap();
+        // Duplicate pb serialization to create unusual but correct protobuf message.
+        payload.append(&mut payload.clone());
+
+        let signed = SignedMessage::new(payload, pub_key, &secret_key);
+
+        let bytes = signed.into_bytes();
+        let message =
+            SignedMessage::from_bytes(bytes.into()).expect("Cannot deserialize signed message");
+        let deserialized_precommit = message
+            .into_verified::<Precommit>()
+            .expect("Failed to check precommit");
+        assert_eq!(precommit_msg, *deserialized_precommit.payload())
     }
 
-    /// Returns round of the message.
-    pub fn round(&self) -> Round {
-        match *self {
-            Consensus::Propose(ref msg) => msg.payload().round(),
-            Consensus::Prevote(ref msg) => msg.payload().round(),
-            Consensus::Precommit(ref msg) => msg.payload().round(),
-        }
-    }
-}
+    #[test]
+    fn test_precommit_serde_correct() {
+        let (pub_key, secret_key) = gen_keypair();
+        let ts = Utc::now();
 
-impl BinaryValue for Message {
-    fn to_bytes(&self) -> Vec<u8> {
-        self.as_raw().to_bytes()
+        let precommit = Verified::from_value(
+            Precommit::new(
+                ValidatorId(123),
+                Height(15),
+                Round(25),
+                crypto::hash(&[1, 2, 3]),
+                crypto::hash(&[3, 2, 1]),
+                ts,
+            ),
+            pub_key,
+            &secret_key,
+        );
+
+        let precommit_json = serde_json::to_string(&precommit).unwrap();
+        let precommit2: Verified<Precommit> = serde_json::from_str(&precommit_json).unwrap();
+        assert_eq!(precommit2, precommit);
     }
 
-    fn from_bytes(value: Cow<'_, [u8]>) -> Result<Self, failure::Error> {
-        let message = SignedMessage::from_bytes(value)?;
-        Self::from_signed(message)
-    }
-}
+    #[test]
+    #[should_panic(expected = "Failed to verify signature.")]
+    fn test_precommit_serde_wrong_signature() {
+        let (pub_key, secret_key) = gen_keypair();
+        let ts = Utc::now();
 
-impl ObjectHash for Message {
-    fn object_hash(&self) -> Hash {
-        self.as_raw().object_hash()
+        let mut precommit = Verified::from_value(
+            Precommit::new(
+                ValidatorId(123),
+                Height(15),
+                Round(25),
+                crypto::hash(&[1, 2, 3]),
+                crypto::hash(&[3, 2, 1]),
+                ts,
+            ),
+            pub_key,
+            &secret_key,
+        );
+        // Break signature.
+        precommit.raw.signature = Signature::zero();
+
+        let precommit_json = serde_json::to_string(&precommit).unwrap();
+        let precommit2: Verified<Precommit> = serde_json::from_str(&precommit_json).unwrap();
+        assert_eq!(precommit2, precommit);
     }
 }
