+++ exonum/src/messages/signed.rs
@@ -20,11 +20,14 @@ use serde::{
     ser::{Serialize, Serializer},
 };
 
-use std::{borrow::Cow, convert::TryFrom};
+use std::{
+    borrow::Cow,
+    convert::{TryFrom, TryInto},
+};
 
 use crate::{
     crypto::{self, Hash, PublicKey, SecretKey},
-    messages::types::{ExonumMessage, SignedMessage},
+    messages::types::SignedMessage,
     proto,
 };
 
@@ -76,8 +79,10 @@ impl_serde_hex_for_binary_value! { SignedMessage }
 /// See module [documentation](index.html#examples) for examples.
 #[derive(Clone, Debug)]
 pub struct Verified<T> {
-    pub(super) raw: SignedMessage,
-    pub(super) inner: T,
+    #[doc(hidden)] // FIXME: how to keep fields private?
+    pub raw: SignedMessage,
+    #[doc(hidden)]
+    pub inner: T,
 }
 
 impl<T> PartialEq for Verified<T> {
@@ -118,18 +123,23 @@ where
     }
 }
 
+/// Message that can be converted into a uniform presentation. The uniformity should be guaranteed
+/// by the implementation.
+pub trait IntoMessage: Sized {
+    /// Container for the message.
+    type Container: BinaryValue + From<Self> + TryInto<Self>;
+}
+
 impl<T> Verified<T>
 where
-    T: TryFrom<SignedMessage> + Into<ExonumMessage> + TryFrom<ExonumMessage>,
+    T: TryFrom<SignedMessage> + IntoMessage,
 {
     /// Signs the specified value and creates a new verified message from it.
     pub fn from_value(inner: T, public_key: PublicKey, secret_key: &SecretKey) -> Self {
-        // Curious trick to avoid clone.
-        // Converts inner to the `ExonumMessage` type to proper serialization.
-        let exonum_msg = inner.into();
-        let raw = SignedMessage::new(exonum_msg.to_bytes(), public_key, secret_key);
+        let container: T::Container = inner.into();
+        let raw = SignedMessage::new(container.to_bytes(), public_key, secret_key);
         // Converts back to the inner type.
-        let inner = if let Ok(inner) = T::try_from(exonum_msg) {
+        let inner: T = if let Ok(inner) = container.try_into() {
             inner
         } else {
             unreachable!("We can safely convert `ExonumMessage` back to the inner type.")
@@ -226,71 +236,9 @@ mod tests {
 
     use super::*;
     use crate::{
-        crypto::{self, Hash},
-        helpers::Height,
-        messages::types::{ExonumMessage, Precommit, Status},
+        crypto,
         runtime::{AnyTx, CallInfo},
     };
-
-    #[test]
-    fn test_verified_from_signed_correct_signature() {
-        let keypair = crypto::gen_keypair();
-
-        let msg = Status {
-            height: Height(0),
-            last_hash: Hash::zero(),
-            pool_size: 0,
-        };
-        let protocol_message = ExonumMessage::from(msg.clone());
-        let signed = SignedMessage::new(protocol_message.clone(), keypair.0, &keypair.1);
-
-        let verified_protocol = signed.clone().into_verified::<ExonumMessage>().unwrap();
-        assert_eq!(verified_protocol.inner, protocol_message);
-
-        let verified_status = signed.clone().into_verified::<Status>().unwrap();
-        assert_eq!(verified_status.inner, msg);
-
-        // Wrong variant
-        let err = signed.into_verified::<Precommit>().unwrap_err();
-        assert_eq!(err.to_string(), "Failed to decode message from payload.");
-    }
-
-    #[test]
-    fn test_verified_from_signed_incorrect_signature() {
-        let keypair = crypto::gen_keypair();
-
-        let msg = Status {
-            height: Height(0),
-            last_hash: Hash::zero(),
-            pool_size: 0,
-        };
-        let protocol_message = ExonumMessage::from(msg.clone());
-        let mut signed = SignedMessage::new(protocol_message.clone(), keypair.0, &keypair.1);
-        // Update author
-        signed.author = crypto::gen_keypair().0;
-        let err = signed.clone().into_verified::<ExonumMessage>().unwrap_err();
-        assert_eq!(err.to_string(), "Failed to verify signature.");
-    }
-
-    #[test]
-    fn test_verified_status_binary_value() {
-        let keypair = crypto::gen_keypair();
-
-        let msg = Verified::from_value(
-            Status {
-                height: Height(0),
-                last_hash: Hash::zero(),
-                pool_size: 0,
-            },
-            keypair.0,
-            &keypair.1,
-        );
-        assert_eq!(msg.object_hash(), msg.as_raw().object_hash());
-
-        let bytes = msg.to_bytes();
-        let msg2 = Verified::<Status>::from_bytes(bytes.into()).unwrap();
-        assert_eq!(msg, msg2);
-    }
 
     #[test]
     fn test_verified_any_tx_binary_value() {
