+++ exonum/src/proto/schema/exonum/proofs.proto
@@ -19,7 +19,7 @@ package exonum;
 option java_package = "com.exonum.core.messages";
 
 import "blockchain.proto";
-import "consensus.proto";
+import "messages.proto";
 import "map_proof.proto";
 
 // Block with its `Precommit` messages.
@@ -33,7 +33,7 @@ message BlockProof {
   // in the block, etc.
   Block block = 1;
   // List of `Precommit` messages for the block.
-  repeated consensus.Verified precommits = 2;
+  repeated messages.Verified precommits = 2;
 }
 
 // Proof of authenticity for a single index within the database.
